On the home page, please change current text (good news/bad news) to:
DECTech is a Colorado School of Mines outreach program designed
especially for girls! Program leaders are female Mines students who are
both passionate about their subjects of study and sharing this passion
with the next generation. Our mission is to engage girls in STEM by
introducing them to various subjects with an interactive, engaging, and
hands-on approach!

Replace the first paragraph on the ABOUT page with this:

Studies show that girls' interest in science/engineering subjects starts
to decline the closer the girls get to middle school. As a result, far
fewer women pursue careers in the various STEM fields than their male
counterparts.

In response to this information, in the Fall of 2012, Dr. Tracy Camp and
her female Computer Science students created a program targeting the
3rd-6th grade age group. This program is designed to foster and continue
girls' interest in the STEM subjects through creative and interactive
activities. The program is very popular with local elementary students,
and has grown to include middle and high school aged girls.

To read more about what female industry leaders and influencers say
about women in the STEM fields, and their ideas on how to get more girls
into STEM, click this link:

http://www.cnn.com/2014/10/27/world/europe/how-to-get-girls/

DECTech is happy to say we satisfy #2, #3, #4, #5, #9 directly … and a
few others indirectly!

(then continue with the MLK information. You may need to adjust the
layout with the two photos)

on the DISCOVER TECH main page:

please remove "Fall 2016" and please replace "To register for this
program, click the REGISTER link in the blue navigation bar above." with
"This program meets four times per semester on School of Mines campus.
It introduces girls to a variety of STEM subjects through fun, hands-on
activities. Registration for Spring 2017 will be open in December 2016.
Keep checking back!"

Remove the fall dates and link to the flyer and keep everything else.

on the EXPLORE TECH main page:

please remove "Fall 2016"

please remove "To register for this program, click the REGISTER link in
the blue navigation bar above."

(keep the rest for now)

More SOON!

Lisa